[UPdate] 플레이어 자힐 기능 추가

+++ Assets/AC7/Script/BSJ/Combat.cs
@@ -117,7 +117,7 @@ public class Combat
         CalcTakeDamage(damage);;
         return true;
     }
-    public void Heal(int v)
+    public void Heal(float v)
     {
         if (_hp < _maxHp)
         {
